requirements.txt
beautifulsoup4>=4.12
google-api-python-client>=2.127
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
python-dateutil>=2.9
streamlit>=1.32
openai>=0.28
